Add people group dashboard support
